Replace Int CrudRepository arg by Long

+++ src/main/java/com/codecademy/goldmedal/repositories/CountryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface CountryRepository extends CrudRepository<Country, Integer> {
+public interface CountryRepository extends CrudRepository<Country, Long> {
   Optional<Country> findCountryByName(String countryName);
   List<Country> getAllByOrderByNameAsc();
   List<Country> getAllByOrderByNameDesc();
